+++ backend/src/user/dto/user.dto.ts
@@ -26,15 +26,22 @@ export class UserDTO {
   isAdmin: boolean;
 
   @Expose()
+  isLdap: boolean;
+
+  ldapDN?: string;
+
+  @Expose()
   totpVerified: boolean;
 
   from(partial: Partial<UserDTO>) {
-    return plainToClass(UserDTO, partial, { excludeExtraneousValues: true });
+    const result = plainToClass(UserDTO, partial, {
+      excludeExtraneousValues: true,
+    });
+    result.isLdap = partial.ldapDN?.length > 0;
+    return result;
   }
 
   fromList(partial: Partial<UserDTO>[]) {
-    return partial.map((part) =>
-      plainToClass(UserDTO, part, { excludeExtraneousValues: true }),
-    );
+    return partial.map((part) => this.from(part));
   }
 }