+++ backend/src/user/user.module.ts
@@ -8,5 +8,6 @@ import { FileModule } from "src/file/file.module";
   imports: [EmailModule, FileModule],
   providers: [UserSevice],
   controllers: [UserController],
+  exports: [UserSevice],
 })
 export class UserModule {}