+++ docs/docusaurus.config.ts
@@ -7,7 +7,7 @@ const config: Config = {
   tagline:
     "Pingvin Share is self-hosted file sharing platform and an alternative for WeTransfer.",
   favicon: "img/pingvinshare.svg",
-  
+
   url: "https://stonith404.github.io",
   baseUrl: "/pingvin-share/",
   organizationName: "stonith404",
@@ -28,7 +28,7 @@ const config: Config = {
         docs: {
           routeBasePath: "/",
           sidebarPath: "./sidebars.ts",
-          editUrl: "https://github.com/stonith404/pingvin-share/docs",
+          editUrl: "https://github.com/stonith404/pingvin-share/edit/main/docs",
         },
         blog: false,
       } satisfies Preset.Options,
@@ -37,6 +37,9 @@ const config: Config = {
 
   themeConfig: {
     image: "img/pingvinshare.svg",
+    colorMode: {
+      respectPrefersColorScheme: true,
+    },
     navbar: {
       title: "Pingvin Share",
       logo: {