+++ frontend/src/i18n/locales.ts
@@ -22,6 +22,9 @@ import ukrainian from "./translations/uk-UA";
 import chineseSimplified from "./translations/zh-CN";
 import chineseTraditional from "./translations/zh-TW";
 import turkish from "./translations/tr-TR";
+import czech from "./translations/cs-CZ";
+import viatnamese from "./translations/vi-VN";
+import bulgarian from "./translations/bg-BG";
 
 export const LOCALES = {
   ENGLISH: {
@@ -143,5 +146,20 @@ export const LOCALES = {
     name: "Türkçe",
     code: "tr-TR",
     messages: turkish,
+  },
+  CZECH: {
+    name: "Čeština",
+    code: "cs-CZ",
+    messages: czech,
+  },
+  VIATNAMESE: {
+    name: "Tiếng Việt",
+    code: "vi-VN",
+    messages: viatnamese,
+  },
+  BULGARIAN: {
+    name: "Български",
+    code: "bg-BG",
+    messages: bulgarian,
   },
 };